+++ tests/backend/version/source/test_env.py
@@ -1,10 +1,7 @@
-from itertools import product
-
 import pytest
 
+from hatch.utils.structures import EnvVars
 from hatchling.version.source.env import EnvSource
-
-DEFAULT_PATTERN_PRODUCTS = list(product(('__version__', 'VERSION', 'version'), ('"', "'"), ('', 'v')))
 
 
 def test_no_variable(isolation):
@@ -21,18 +18,17 @@ def test_variable_not_string(isolation):
         source.get_version_data()
 
 
-def test_variable_not_available(isolation, monkeypatch):
+def test_variable_not_available(isolation):
     source = EnvSource(str(isolation), {'variable': 'ENV_VERSION'})
 
-    monkeypatch.delenv('ENV_VERSION', raising=False)
-
-    with pytest.raises(RuntimeError, match='variable `ENV_VERSION` is not set'):
+    with EnvVars(exclude=['ENV_VERSION']), pytest.raises(
+        RuntimeError, match='environment variable `ENV_VERSION` is not set'
+    ):
         source.get_version_data()
 
 
-def test_variable_contains_version(isolation, monkeypatch):
+def test_variable_contains_version(isolation):
     source = EnvSource(str(isolation), {'variable': 'ENV_VERSION'})
 
-    monkeypatch.setenv('ENV_VERSION', '0.0.1')
-
-    assert source.get_version_data()['version'] == '0.0.1'
+    with EnvVars({'ENV_VERSION': '0.0.1'}):
+        assert source.get_version_data()['version'] == '0.0.1'