+++ docs/history.md
@@ -26,7 +26,6 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 - Support Bash on Windows for the `shell` command
 - The `setuptools` migration script no longer modifies the formatting of existing `pyproject.toml` configuration
 - Bump the minimum supported version of Hatchling to 1.10.0
-- Add `env` as version number source to retrieve the version from an environment variable.
 
 ***Fixed:***
 
@@ -182,6 +181,10 @@ This is the first stable release of Hatch v1, a complete rewrite. Enjoy!
 ## **Hatchling**
 
 ### Unreleased
+
+***Added:***
+
+- Add `env` version source to retrieve the version from an environment variable
 
 ***Fixed:***